crypto: msm: Fix failure to process AES-XTS 256 key
For AES-XTS 256 mode, key size is 64 bytes (twice the
AES key size of 32 bytes).  The driver  only checks for
AES key size and not specifically for xts key size.
The driver checks for AES key size and return a failure
for AES XTS 256 where key size is 64 bytes

Fix is to check for valid keys for XTS algorithm.

Change-Id: Ia7bafa5b5c13eab018a2e1d729a82c8d329a092e
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>