Add the BFS CPU Scheduler
This incorporates the 0.413 version of the BFS CPU Scheduler created by
Con Kolivas. _thalamus was kind of enough to backport some features from
the 0.420 version of the scheduler as well. In addition, _thalamus also
implemented a tweak that will save several CPU cycles.

Modified to fit with ancora - fix some compilation error